backend/requirement.txt
Django==4.0.5
django-admin-safe-login==0.3.4
django-cors-headers==3.11.0
django-oauth-toolkit==2.0.0
django-ranged-response==0.2.0
django-rest-framework-social-oauth2==1.1.0
django-static-jquery3==5.1.0
django-storages==1.12.3
djangorestframework==3.13.1
drf-yasg==1.20.0
facebook-sdk==3.1.0
google-api-core==2.8.1
google-auth==2.6.6
google-cloud==0.34.0
google-cloud-vision==2.7.2
oauthlib==3.2.0
psycopg2-binary==2.9.3
requests-oauthlib==1.3.1